refactor: remove marshalUriFromSapi function
Marked for removal for 3.0.0, so this patch removes it.
It also removes tests in `ServerRequestFactoryTest` that exercised it, as these had been largely moved to the `UriFactoryTest` already.
One test method was found that had not, and was moved to that factory.
