+++ composer.json
@@ -60,7 +60,6 @@
             "src/functions/marshal_headers_from_sapi.php",
             "src/functions/marshal_method_from_sapi.php",
             "src/functions/marshal_protocol_version_from_sapi.php",
-            "src/functions/marshal_uri_from_sapi.php",
             "src/functions/normalize_server.php",
             "src/functions/normalize_uploaded_files.php",
             "src/functions/parse_cookie_header.php"
