+++ test/ServerRequestFactoryTest.php
@@ -9,21 +9,15 @@ use Laminas\Diactoros\ServerRequestFactory;
 use Laminas\Diactoros\ServerRequestFilter\DoNotFilter;
 use Laminas\Diactoros\ServerRequestFilter\FilterServerRequestInterface;
 use Laminas\Diactoros\UploadedFile;
-use Laminas\Diactoros\Uri;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 use UnexpectedValueException;
 
-use function array_shift;
 use function Laminas\Diactoros\marshalHeadersFromSapi;
 use function Laminas\Diactoros\marshalProtocolVersionFromSapi;
-use function Laminas\Diactoros\marshalUriFromSapi;
 use function Laminas\Diactoros\normalizeServer;
 use function Laminas\Diactoros\normalizeUploadedFiles;
-use function sprintf;
-use function str_contains;
 use function str_replace;
-use function strtolower;
 
 final class ServerRequestFactoryTest extends TestCase
 {
@@ -93,296 +87,6 @@ final class ServerRequestFactoryTest extends TestCase
         ];
 
         $this->assertEquals($expected, marshalHeadersFromSapi($server));
-    }
-
-    public function testMarshalRequestUriUsesIISUnencodedUrlValueIfPresentAndUrlWasRewritten(): void
-    {
-        $server = [
-            'IIS_WasUrlRewritten' => '1',
-            'UNENCODED_URL'       => '/foo/bar',
-        ];
-
-        $uri = marshalUriFromSapi($server, []);
-
-        $this->assertSame($server['UNENCODED_URL'], $uri->getPath());
-    }
-
-    public function testMarshalRequestUriStripsSchemeHostAndPortInformationWhenPresent(): void
-    {
-        $server = [
-            'REQUEST_URI' => 'http://example.com:8000/foo/bar',
-        ];
-
-        $uri = marshalUriFromSapi($server, []);
-
-        $this->assertSame('/foo/bar', $uri->getPath());
-    }
-
-    public function testMarshalRequestUriUsesOrigPathInfoIfPresent(): void
-    {
-        $server = [
-            'ORIG_PATH_INFO' => '/foo/bar',
-        ];
-
-        $uri = marshalUriFromSapi($server, []);
-
-        $this->assertSame('/foo/bar', $uri->getPath());
-    }
-
-    public function testMarshalRequestUriFallsBackToRoot(): void
-    {
-        $server = [];
-
-        $uri = marshalUriFromSapi($server, []);
-
-        $this->assertSame('/', $uri->getPath());
-    }
-
-    public function testMarshalHostAndPortUsesHostHeaderWhenPresent(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withMethod('GET');
-        $request = $request->withHeader('Host', 'example.com');
-
-        $uri = marshalUriFromSapi([], $request->getHeaders());
-
-        $this->assertSame('example.com', $uri->getHost());
-        $this->assertNull($uri->getPort());
-    }
-
-    public function testMarshalHostAndPortWillDetectPortInHostHeaderWhenPresent(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com:8000/'));
-        $request = $request->withMethod('GET');
-        $request = $request->withHeader('Host', 'example.com:8000');
-
-        $uri = marshalUriFromSapi([], $request->getHeaders());
-
-        $this->assertSame('example.com', $uri->getHost());
-        $this->assertSame(8000, $uri->getPort());
-    }
-
-    public function testMarshalHostAndPortReturnsEmptyValuesIfNoHostHeaderAndNoServerName(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri());
-
-        $uri = marshalUriFromSapi([], $request->getHeaders());
-
-        $this->assertSame('', $uri->getHost());
-        $this->assertNull($uri->getPort());
-    }
-
-    public function testMarshalHostAndPortReturnsServerNameForHostWhenPresent(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-
-        $server = [
-            'SERVER_NAME' => 'example.com',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertSame('example.com', $uri->getHost());
-        $this->assertNull($uri->getPort());
-    }
-
-    public function testMarshalHostAndPortReturnsServerPortForPortWhenPresentWithServerName(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri());
-
-        $server = [
-            'SERVER_NAME' => 'example.com',
-            'SERVER_PORT' => 8000,
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertSame('example.com', $uri->getHost());
-        $this->assertSame(8000, $uri->getPort());
-    }
-
-    public function testMarshalHostAndPortReturnsServerNameForHostIfServerAddrPresentButHostIsNotIpv6Address(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-
-        $server = [
-            'SERVER_ADDR' => '127.0.0.1',
-            'SERVER_NAME' => 'example.com',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertSame('example.com', $uri->getHost());
-    }
-
-    public function testMarshalHostAndPortReturnsServerAddrForHostIfPresentAndHostIsIpv6Address(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri());
-
-        $server = [
-            'SERVER_ADDR' => 'FE80::0202:B3FF:FE1E:8329',
-            'SERVER_NAME' => '[FE80::0202:B3FF:FE1E:8329]',
-            'SERVER_PORT' => 8000,
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertSame(strtolower('[FE80::0202:B3FF:FE1E:8329]'), $uri->getHost());
-        $this->assertSame(8000, $uri->getPort());
-    }
-
-    public function testMarshalHostAndPortWillDetectPortInIpv6StyleHost(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri());
-
-        $server = [
-            'SERVER_ADDR' => 'FE80::0202:B3FF:FE1E:8329',
-            'SERVER_NAME' => '[FE80::0202:B3FF:FE1E:8329:80]',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertSame(strtolower('[FE80::0202:B3FF:FE1E:8329]'), $uri->getHost());
-        $this->assertNull($uri->getPort());
-    }
-
-    /** @return non-empty-array<non-empty-string, array{non-empty-string}> */
-    public function httpsParamProvider(): array
-    {
-        return [
-            'lowercase' => ['https'],
-            'uppercase' => ['HTTPS'],
-        ];
-    }
-
-    /**
-     * @dataProvider httpsParamProvider
-     * @param non-empty-string $param
-     */
-    public function testMarshalUriDetectsHttpsSchemeFromServerValue(string $param): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withHeader('Host', 'example.com');
-
-        $server = [
-            $param => 'on',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertInstanceOf(Uri::class, $uri);
-        $this->assertSame('https', $uri->getScheme());
-    }
-
-    /** @return iterable<string, array{non-empty-string, 'off'|'OFF'}> */
-    public function httpsDisableParamProvider(): iterable
-    {
-        foreach ($this->httpsParamProvider() as $key => $data) {
-            $param = array_shift($data);
-            foreach (['lowercase-off', 'uppercase-off'] as $type) {
-                $key   = sprintf('%s-%s', $key, $type);
-                $value = str_contains($type, 'lowercase') ? 'off' : 'OFF';
-                yield $key => [$param, $value];
-            }
-        }
-    }
-
-    /**
-     * @dataProvider httpsDisableParamProvider
-     * @param non-empty-string $param
-     * @param 'off'|'OFF' $value
-     */
-    public function testMarshalUriUsesHttpSchemeIfHttpsServerValueEqualsOff(string $param, string $value): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withHeader('Host', 'example.com');
-
-        $server = [
-            $param => $value,
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertInstanceOf(Uri::class, $uri);
-        $this->assertSame('http', $uri->getScheme());
-    }
-
-    /**
-     * @dataProvider httpsParamProvider
-     * @param non-empty-string $xForwardedProto
-     */
-    public function testMarshalUriDetectsHttpsSchemeFromXForwardedProtoValue(string $xForwardedProto): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withHeader('Host', 'example.com');
-        $request = $request->withHeader('X-Forwarded-Proto', $xForwardedProto);
-
-        $server = [];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertInstanceOf(Uri::class, $uri);
-        $this->assertSame('https', $uri->getScheme());
-    }
-
-    public function testMarshalUriStripsQueryStringFromRequestUri(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withHeader('Host', 'example.com');
-
-        $server = [
-            'REQUEST_URI' => '/foo/bar?foo=bar',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertInstanceOf(Uri::class, $uri);
-        $this->assertSame('/foo/bar', $uri->getPath());
-    }
-
-    public function testMarshalUriInjectsQueryStringFromServer(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withHeader('Host', 'example.com');
-
-        $server = [
-            'REQUEST_URI'  => '/foo/bar?foo=bar',
-            'QUERY_STRING' => 'bar=baz',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertInstanceOf(Uri::class, $uri);
-        $this->assertSame('bar=baz', $uri->getQuery());
-    }
-
-    public function testMarshalUriInjectsFragmentFromServer(): void
-    {
-        $request = new ServerRequest();
-        $request = $request->withUri(new Uri('http://example.com/'));
-        $request = $request->withHeader('Host', 'example.com');
-
-        $server = [
-            'REQUEST_URI' => '/foo/bar#foo',
-        ];
-
-        $uri = marshalUriFromSapi($server, $request->getHeaders());
-
-        $this->assertInstanceOf(Uri::class, $uri);
-        $this->assertSame('foo', $uri->getFragment());
     }
 
     public function testCanCreateServerRequestViaFromGlobalsMethod(): void
@@ -601,19 +305,6 @@ final class ServerRequestFactoryTest extends TestCase
             'HTTP/1.1' => ['HTTP/1.1', '1.1'],
             'HTTP/2'   => ['HTTP/2', '2'],
         ];
-    }
-
-    public function testMarshalRequestUriPrefersRequestUriServerParamWhenXOriginalUrlButNoXRewriteUrlPresent(): void
-    {
-        $headers = [
-            'X-Original-URL' => '/hijack-attempt',
-        ];
-        $server  = [
-            'REQUEST_URI' => 'https://example.com/requested/path',
-        ];
-
-        $uri = marshalUriFromSapi($server, $headers);
-        $this->assertSame('/requested/path', $uri->getPath());
     }
 
     public function testServerRequestFactoryHasAWritableEmptyBody(): void
